[Upstream] Code cleanup and small optimizations in RecastFilter.cpp rcFilterLedgeSpans (https://github.com/ikpil/DotRecast/pull/29)
- https://github.com/recastnavigation/recastnavigation/commit/3e94c3b6fcb5361bc30e87cacfa0ff46bde37a2c

* Code cleanup and minor refactor in RecastFilter.cpp rcFilterLedgeSpans

Because span.smax is always > 0, bot > 0 as well, and (-walkableClimb - bot) is always < -walkableClimb. Furthermore, as long as minNeighborHeight < -walkableClimb' at least once, there is no need to continue the traversal.

* Code cleanup and minor refactor in RecastFilter.cpp rcFilterLedgeSpans

Because span.smax is always > 0, bot > 0 as well, and (-walkableClimb - bot) is always < -walkableClimb. Furthermore, as long as minNeighborHeight < -walkableClimb' at least once, there is no need to continue the traversal.

* Update RecastFilter.cpp